备份3：修复 P1 运行逻辑 bug（Kalman/PDOA/4G/WiFi/海拔）
- algorithm.c: pit 公式整数除法改浮点（(double)pdoa1 / 6434.0）
- Kalman 滤波器改为每基站独立状态 kf_twr[MAX_ANCHOR_LIST_SIZE]，update 带 src 索引和越界保护
- 4Gdrv.c: static uint8_t i 改 int（-1 返回值不再被截断）；QISEND 发送长度由 sizeof(buf)=20 改为精确 17 字节
- WIFIdrv.c: e103_send_cmd 补 m++ 使重试退出条件可触发；i 改 int 且成功判断改为 i == 0
- CM7 main.c: g_SkpAltitude extern 声明 uint32_t 改 uint16_t（与定义和 frame.alt 一致）
- 验证：CM4、CM7 均编译链接通过

+++ CM7/Core/Src/main.c
@@ -82,7 +82,7 @@ extern uint8_t  ec20_reboot_num;
 
 extern uint32_t g_Pressure;
 extern uint32_t g_MeasureFreq;
-extern uint32_t g_SkpAltitude;
+extern uint16_t g_SkpAltitude;
 extern int16_t  g_EulerAngles[3];//X,Y,Z
 extern int16_t  g_Acceleration[3];//X,Y,Z
 extern int16_t  g_MagneticField[3];//X,Y,Z
